move draw functions to lib/draw
