+++ packages/components/src/lib/draw/index.ts
@@ -1,9 +1,4 @@
 import * as d3 from "d3";
-import range from "lodash/range.js";
-
-import { SqLambda, SqValue } from "@quri/squiggle-lang";
-
-import { FunctionChartSettings } from "../components/FunctionChart/index.js";
 
 const axisColor = "rgba(114, 125, 147, 0.1)";
 export const labelColor = "rgb(114, 125, 147)";
@@ -12,62 +7,6 @@ export const primaryColor = "#4c78a8"; // for lines and areas
 const labelFont = "10px sans-serif";
 const xLabelOffset = 6;
 const yLabelOffset = 6;
-
-function rangeByCount(start: number, stop: number, count: number) {
-  const step = (stop - start) / (count - 1);
-  const items = range(start, stop, step);
-  const result = items.concat([stop]);
-  return result;
-}
-
-type Subvalue<T> = T extends SqValue["tag"]
-  ? Extract<SqValue, { tag: T }>
-  : never;
-
-export function getFunctionImage<T extends Exclude<SqValue["tag"], "Void">>({
-  settings,
-  fn,
-  valueType,
-}: {
-  settings: FunctionChartSettings;
-  fn: SqLambda;
-  valueType: T;
-}) {
-  const chartPointsToRender = rangeByCount(
-    settings.start,
-    settings.stop,
-    settings.count
-  );
-
-  let functionImage: {
-    x: number;
-    y: Subvalue<T>["value"];
-  }[] = [];
-  let errors: { x: number; value: string }[] = [];
-
-  for (const x of chartPointsToRender) {
-    const result = fn.call([x]);
-    if (result.ok) {
-      if (result.value.tag === valueType) {
-        functionImage.push({
-          x,
-          // This is sketchy, I'm not sure why the type check passes (it's not because of `if` guard above), might be a Typescript bug.
-          // The result should be correct, though.
-          y: result.value.value,
-        });
-      } else {
-        errors.push({
-          x,
-          value: `This component expected outputs of type ${valueType}`,
-        });
-      }
-    } else {
-      errors.push({ x, value: result.value.toString() });
-    }
-  }
-
-  return { errors, functionImage };
-}
 
 export type Padding = {
   top: number;
@@ -211,7 +150,7 @@ export function drawAxes({
         context.lineTo(x - tickSize, y);
         context.stroke();
       }
-      context.fillText(yTickFormat(d), x - 6, y);
+      context.fillText(yTickFormat(d), x - yLabelOffset, y);
     });
     context.restore();
   }
